+++ modules/mesh-task/variables.tf
@@ -45,6 +45,24 @@ variable "volumes" {
   default     = []
 }
 
+variable "task_role" {
+  description = "ECS task role to include in the task definition. If not provided, a role is created."
+  type = object({
+    id  = string
+    arn = string
+  })
+  default = null
+}
+
+variable "execution_role" {
+  description = "ECS execution role to include in the task definition. If not provided, a role is created."
+  type = object({
+    id  = string
+    arn = string
+  })
+  default = null
+}
+
 variable "additional_task_role_policies" {
   description = "List of additional policy ARNs to attach to the task role."
   type        = list(string)